requirements.txt
openai
PyMuPDF
python-dotenv
tiktoken
faiss-cpu
numpy